Removed percentage-based IOPS field. Changed iops_min/max to assured/throttle and unified the data types.

+++ source/java/src/com/formationds/om/webkit/rest/CreateVolume.java
@@ -173,9 +173,9 @@ public class CreateVolume
           Thread.sleep( 200 );
           SetVolumeQosParams.setVolumeQos( legacyConfigPath,
                                            volume.getName(),
-                                           ( int ) volume.getSla(),
+                                           volume.getSla(),
                                            volume.getPriority(),
-                                           ( int ) volume.getLimit(),
+                                           volume.getLimit(),
                                            volume.getCommit_log_retention(),
                                            volume.getMediaPolicy() );
